# src/quads/server/blueprints/hosts.py
import asyncio

from datetime import datetime

from flask import Blueprint, Response, jsonify, make_response, request
from sqlalchemy import inspect

from quads.config import Config
from quads.plugins.dispatchers.provisioner import get_images
from quads.server.blueprints import check_access
from quads.server.dao.baseDao import BaseDao, EntryNotFound, InvalidArgument
from quads.server.dao.cloud import CloudDao
from quads.server.dao.host import HostDao
from quads.server.models import Host, db

host_bp = Blueprint("hosts", __name__)


@host_bp.route("/")
def get_hosts() -> Response:
    if request.args:
        try:
            _hosts = HostDao.filter_hosts_dict(request.args)
        except (EntryNotFound, InvalidArgument) as ex:
            response = {
                "status_code": 400,
                "error": "Bad Request",
                "message": str(ex),
            }
            return make_response(jsonify(response), 400)

    else:
        _hosts = HostDao.get_hosts()

    _hosts_dict = {}
    if _hosts:
        if isinstance(_hosts[0], Host):
            _hosts_dict = [_host.as_dict() for _host in _hosts]
        else:
            _hosts_dict = {key: value for key, value in _hosts}

    return make_response(jsonify(_hosts_dict), 200)


@host_bp.route("/availability_summary")
def get_availability_summary() -> Response:
    _params = request.args.to_dict()
    _format = "%Y-%m-%dT%H:%M"
    try:
        now = datetime.strptime(_params["now"], _format)
        two_week_start = datetime.strptime(_params["two_week_start"], _format)
        two_week_end = datetime.strptime(_params["two_week_end"], _format)
        four_week_end = datetime.strptime(_params["four_week_end"], _format)
    except (KeyError, ValueError) as ex:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": f"Invalid or missing date parameter: {ex}",
        }
        return make_response(jsonify(response), 400)

    results = HostDao.get_availability_summary(now, two_week_start, two_week_end, four_week_end)
    summary = [
        {
            "model": row.model,
            "total": row.total,
            "scheduled": int(row.scheduled),
            "avail_2w": int(row.avail_2w),
            "avail_4w": int(row.avail_4w),
        }
        for row in results
    ]
    return jsonify(summary)


@host_bp.route("/<hostname>")
def get_host(hostname: str) -> Response:
    _host = HostDao.get_host(hostname)
    if not _host:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": f"Host not found: {hostname}",
        }
        return make_response(jsonify(response), 400)
    return jsonify(_host.as_dict())


@host_bp.route("/<hostname>/memory")
def get_host_memory(hostname: str) -> Response:
    _host = HostDao.get_host(hostname)
    if not _host:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": f"Host not found: {hostname}",
        }
        return make_response(jsonify(response), 400)
    return jsonify([_memory.as_dict() for _memory in _host.memory])


@host_bp.route("/<hostname>/processors")
def get_host_processors(hostname: str) -> Response:
    _host = HostDao.get_host(hostname)
    if not _host:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": f"Host not found: {hostname}",
        }
        return make_response(jsonify(response), 400)
    return jsonify([_processors.as_dict() for _processors in _host.processors])


@host_bp.route("/<hostname>/disks")
def get_host_disks(hostname: str) -> Response:
    _host = HostDao.get_host(hostname)
    if not _host:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": f"Host not found: {hostname}",
        }
        return make_response(jsonify(response), 400)
    return jsonify([_disks.as_dict() for _disks in _host.disks])


@host_bp.route("/<hostname>/interfaces")
def get_host_interfaces(hostname: str) -> Response:
    _host = HostDao.get_host(hostname)
    if not _host:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": f"Host not found: {hostname}",
        }
        return make_response(jsonify(response), 400)
    return jsonify([_interface.as_dict() for _interface in _host.interfaces])


@host_bp.route("/<hostname>", methods=["PATCH"])
@check_access(["admin"])
def update_host(hostname: str) -> Response:
    data = request.get_json()

    _host = HostDao.get_host(hostname)
    if not _host:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": f"Host not found: {hostname}",
        }
        return make_response(jsonify(response), 400)

    obj_attrs = inspect(Host).mapper.attrs
    update_fields = {}
    for attr in obj_attrs:
        value = data.get(attr.key)
        if value is not None:
            if attr.key.lower() in ["cloud", "default_cloud"]:
                _cloud = CloudDao.get_cloud(value)
                if not _cloud:
                    response = {
                        "status_code": 400,
                        "error": "Bad Request",
                        "message": f"Cloud not found: {value}",
                    }
                    return make_response(jsonify(response), 400)

                value = _cloud
            if attr.key.lower() == "model":
                _host.model = value.upper()
            if type(value) is str:
                mapping = {"true": True, "false": False, "none": None}
                if value.lower() in mapping:
                    value = mapping[value.lower()]
            update_fields[attr.key] = value

    for key, value in update_fields.items():
        setattr(_host, key, value)

    BaseDao.safe_commit()

    return jsonify(_host.as_dict())


@host_bp.route("/", methods=["POST"])
@check_access(["admin"])
def create_host() -> Response:
    data = request.get_json()
    hostname = data.get("name")
    model = data.get("model")
    default_cloud = data.get("default_cloud")
    host_type = data.get("host_type")
    can_self_schedule = data.get("can_self_schedule")
    overcloud = data.get("overcloud")
    rack = data.get("rack")
    uloc = data.get("uloc")
    blade = data.get("blade")
    bootmode = data.get("bootmode")

    if not model:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": "Missing argument: model",
        }
        return make_response(jsonify(response), 400)
    else:
        if model.upper() not in Config["models"]:
            response = {
                "status_code": 400,
                "error": "Bad Request",
                "message": f"Model {model} does not seem to be part of the defined models on quads.yml",
            }
            return make_response(jsonify(response), 400)

    if not default_cloud:
        default_cloud = Config["spare_pool_name"]

    if not hostname:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": "Missing argument: name",
        }
        return make_response(jsonify(response), 400)

    if not rack:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": "Missing argument: rack",
        }
        return make_response(jsonify(response), 400)

    if not uloc:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": "Missing argument: uloc",
        }
        return make_response(jsonify(response), 400)

    _host = HostDao.get_host(hostname)
    if _host:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": f"Host {hostname} already exists",
        }
        return make_response(jsonify(response), 400)

    if not host_type:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": "Missing argument: host_type",
        }
        return make_response(jsonify(response), 400)

    if default_cloud:
        _default_cloud = CloudDao.get_cloud(default_cloud)
        if not _default_cloud:
            response = {
                "status_code": 400,
                "error": "Bad Request",
                "message": f"Default Cloud not found: {default_cloud}",
            }
            return make_response(jsonify(response), 400)

    else:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": "Missing argument: default_cloud",
        }
        return make_response(jsonify(response), 400)

    _host_obj = Host(
        name=hostname,
        model=model.upper(),
        host_type=host_type,
        can_self_schedule=can_self_schedule,
        overcloud=overcloud,
        rack=rack,
        uloc=uloc,
        blade=blade,
        bootmode=bootmode,
        cloud=_default_cloud,
        default_cloud=_default_cloud,
    )
    db.session.add(_host_obj)
    BaseDao.safe_commit()
    return make_response(jsonify(_host_obj.as_dict()), 201)


@host_bp.route("/<hostname>", methods=["DELETE"])
@check_access(["admin"])
def delete_host(hostname: str) -> Response:
    _host = HostDao.get_host(hostname)
    if not _host:
        response = {
            "status_code": 400,
            "error": "Bad Request",
            "message": f"Host not found: {hostname}",
        }
        return make_response(jsonify(response), 400)

    HostDao.remove_host(hostname)
    response = {
        "status_code": 200,
        "message": "Host deleted",
    }
    return jsonify(response)


@host_bp.route("/os_list", methods=["GET"])
def get_list_os() -> Response:
    available_os_list = asyncio.run(get_images())
    headers = ["Id", "Title", "Release Name", "Family"]
    formatted_output = [
        {header: available_os[header.lower().replace(" ", "_")] for header in headers}
        for available_os in available_os_list
    ]
    return make_response(jsonify(formatted_output), 200)