# src/quads/server/swagger.yaml
openapi: 3.0.0
info:
  version: '2.2.6'
  title: 'QUADS'
  description: |-
    QUADS automates the future scheduling, end-to-end provisioning and delivery of bare-metal servers and networks.

    This is it's REST API implementation.

    Some useful links:
    - [Quads repository](https://github.com/quadsproject/quads)
    - [Quads blog](https://quads.dev)
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: 'https://quads.example.com/api/v3/'
...
paths:
...
  /hosts/:
    get:
      summary: Returns a list with all the defined hosts
      tags:
        - Hosts
      parameters:
        - in: query
          name: name
          schema:
            type: string
          description: Filter hosts by name.
        - in: query
          name: model
          schema:
            type: string
          description: Filter hosts by model name.
        - in: query
          name: host_type
          schema:
            type: string
          description: Filter hosts by type.
        - in: query
          name: build
          schema:
            type: boolean
          description: Filter by build status.
        - in: query
          name: validated
          schema:
            type: boolean
          description: Filter by validated status.
        - in: query
          name: switch_config_applied
          schema:
            type: boolean
          description: Filter by switch config applied status.
        - in: query
          name: broken
          schema:
            type: boolean
          description: Filter by broken status.
        - in: query
          name: retired
          schema:
            type: boolean
          description: Filter by retired status.
        - in: query
          name: last_build
          schema:
            type: string
          description: Filter by last build date.
        - in: query
          name: created_on
          schema:
            type: string
          description: Filter by created on date.
      responses:
        '200':
          description: Host name
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Host'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add a new host
      tags:
        - Hosts
      requestBody:
        description: Body request with *host* details
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - model
                - host_type
                - rack
                - uloc
              properties:
                name:
                  type: string
                  example: host.example.com
                model:
                  type: string
                  example: r640
                host_type:
                  type: string
                  example: vendor
                rack:
                  type: string
                  example: f01
                uloc:
                  type: string
                  example: h01
                blade:
                  type: string
                  example: b01
                default_cloud:
                  type: string
                  example: cloud01
                can_self_schedule:
                  type: boolean
                  example: false
                overcloud:
                  type: boolean
                  example: false
                bootmode:
                  type: string
                  example: Bios
          text/plain:
            schema:
              type: string
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: [ ]
  /hosts/availability_summary:
    get:
      summary: Returns host availability summary over two and four week windows
      tags:
        - Hosts
      parameters:
        - name: now
          in: query
          description: Current date reference point
          required: true
          schema:
            type: string
        - name: two_week_start
          in: query
          description: Start of the two week reporting window
          required: true
          schema:
            type: string
        - name: two_week_end
          in: query
          description: End of the two week reporting window
          required: true
          schema:
            type: string
        - name: four_week_end
          in: query
          description: End of the four week reporting window
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Host availability summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Host'
        '400':
          description: Bad Request - missing or invalid dates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /hosts/{hostName}:
    get:
      summary: Returns a host by name
      tags:
        - Hosts
      parameters:
        - name: hostName
          in: path
          description: Host name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Host name
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update an existing host
      tags:
        - Hosts
      requestBody:
        description: Body request with *host* details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostClean'
      parameters:
        - name: hostName
          in: path
          description: Host name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Host name
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: [ ]

    delete:
      summary: Delete host by host name
      tags:
        - Hosts
      parameters:
        - name: hostName
          in: path
          description: Hostname
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted Host
      security:
        - BearerAuth: [ ]
  /hosts/{hostName}/memory:
    get:
      summary: Returns a host memory by host name
      tags:
        - Hosts
      parameters:
        - name: hostName
          in: path
          description: Host name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Memory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Memory'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /hosts/{hostName}/processors:
    get:
      summary: Returns a host processors by host name
      tags:
        - Hosts
      parameters:
        - name: hostName
          in: path
          description: Host name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Processors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Processor'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /hosts/{hostName}/disks:
    get:
      summary: Returns a host disks by host name
      tags:
        - Hosts
      parameters:
        - name: hostName
          in: path
          description: Host name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Disks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Disk'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /hosts/{hostName}/interfaces:
    get:
      summary: Returns a host interfaces by host name
      tags:
        - Hosts
      parameters:
        - name: hostName
          in: path
          description: Host name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Interfaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Interface'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /hosts/os_list:
    get:
      summary: Returns a list of available OS images
      tags:
        - Hosts
      responses:
        '200':
          description: Available OS images
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    Id:
                      type: string
                    Title:
                      type: string
                    Release Name:
                      type: string
                    Family:
                      type: string
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Disk:
      type: object
      properties:
        disk_id:
          type: integer
          example: 1
        disk_type:
          type: string
          example: nvme
        size_gb:
          type: integer
          example: 2000
        count:
          type: integer
          example: 10
        host_id:
          type: integer
          example: 10
    Memory:
      type: object
      required:
        - size_gb
        - handle
        - host_id
      properties:
        id:
          type: integer
          example: 12
        size_gb:
          type: integer
          example: 2000
        handle:
          type: string
          example: MMD
        host_id:
          type: integer
          example: 10
    Processor:
      type: object
      properties:
        id:
          type: integer
          example: 12
        handle:
          type: string
          example: CoreI7
        vendor:
          type: string
          example: Intel
        product:
          type: string
          example: i7
        cores:
          type: integer
          example: 20
        threads:
          type: integer
          example: 20
        processor_type:
          type: string
          example: x86
        host_id:
          type: integer
          example: 10
    Interface:
      type: object
      properties:
        id:
          type: integer
          example: 12
        name:
          type: string
          example: em1
        bios_id:
          type: string
          example: nic1
        mac_address:
          type: string
          example: aa:00:bb:11:cc:22
        switch_ip:
          type: string
          example: 10.1.1.18
        switch_port:
          type: string
          example: xt-0-0/1
        speed:
          type: integer
          example: 1000
        vendor:
          type: string
          example: Intel
        pxe_boot:
          type: boolean
          example: true
        maintenance:
          type: boolean
          example: false
        host_id:
          type: integer
          example: 10
    HostClean:
      type: object
      properties:
        name:
          type: string
          example: host.example.com
        model:
          type: string
          example: r640
        host_type:
          type: string
          example: vendor
        build:
          type: boolean
          example: true
        validated:
          type: boolean
          example: true
        switch_config_applied:
          type: boolean
          example: true
        broken:
          type: boolean
          example: true
        retired:
          type: boolean
          example: true
        can_self_schedule:
          type: boolean
          example: false
        overcloud:
          type: boolean
          example: false
        rack:
          type: string
          example: A1
        uloc:
          type: string
          example: U10
        blade:
          type: string
          example: blade1
        bootmode:
          type: string
          example: uefi
        last_build:
          type: string
          example: 2022-02-02
        cloud_id:
          type: integer
          example: 12
        default_cloud_id:
          type: integer
          example: 1
    Host:
      type: object
      properties:
        id:
          type: integer
          example: 12
        name:
          type: string
          example: host.example.com
        model:
          type: string
          example: r640
        host_type:
          type: string
          example: vendor
        build:
          type: boolean
          example: true
        validated:
          type: boolean
          example: true
        switch_config_applied:
          type: boolean
          example: true
        broken:
          type: boolean
          example: true
        retired:
          type: boolean
          example: true
        can_self_schedule:
          type: boolean
          example: false
        overcloud:
          type: boolean
          example: false
        rack:
          type: string
          example: A1
        uloc:
          type: string
          example: U10
        blade:
          type: string
          example: blade1
        bootmode:
          type: string
          example: uefi
        last_build:
          type: string
          example: 2022-02-02
        cloud_id:
          type: integer
          example: 12
        default_cloud_id:
          type: integer
          example: 1
        interfaces:
          type: array
          items:
            $ref: '#/components/schemas/Interface'
        disks:
          type: array
          items:
            $ref: '#/components/schemas/Disk'
        memory:
          type: array
          items:
            $ref: '#/components/schemas/Memory'
        processors:
          type: array
          items:
            $ref: '#/components/schemas/Processor'
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 400
          format: int32
        message:
          type: string
          example: Something went horribly wrong.
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer